Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
google-generativeai==0.3.0
pypdf==3.17.4
faiss-cpu==1.7.4
python-dotenv==1.0.0
langchain-huggingface==0.0.3
tiktoken==0.5.1
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pymongo==4.5.0
argon2-cffi==23.1.0
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
python-pptx==0.6.23
pillow==10.1.0
pytesseract==0.3.10